requirements.txt
django==2.2.14
django-cleanup==5.1.0
django-ckeditor==6.0.0
django-crispy-forms==1.7.2
django-js-asset==1.2.2
